Split domain slider styles into multiple modules
