+++ src/h5web/toolbar/controls/DomainSlider/Track.module.css
@@ -1,0 +1,20 @@
+.track {
+  position: relative;
+  flex: 1 1 0%;
+  align-self: center;
+  height: 0.5rem;
+  margin: 0 calc(var(--thumb-size) / 2); /* make track end in middle of thumbs */
+  background-color: var(--secondary-dark-15);
+  box-shadow: var(--btn-shadow-pressed) var(--btn-shadow-color);
+  border-radius: 0.25rem;
+  overflow: hidden; /* crop data track to root track's border radius */
+}
+
+.dataTrack {
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  background-color: var(--secondary-dark);
+  box-shadow: 0 -2px 3px inset var(--secondary-darker);
+  border-radius: 2px;
+}
